requirements.txt
annotated-types==0.7.0
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
docstring_parser==0.16
firebase-admin==6.5.0
Flask==3.0.3
google-api-core==2.19.1
google-api-python-client==2.135.0
google-auth==2.32.0
google-cloud-aiplatform==1.60.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-resource-manager==1.12.4
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.65.1
grpcio-status==1.62.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
msgpack==1.0.8
numpy==2.0.1
opencv-python==4.10.0.84
packaging==24.1
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.0
shapely==2.0.5
six==1.16.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.2.2
vertexai==1.60.0
Werkzeug==3.0.3
